confluent-rss-newsbot/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.5.3
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
kafka==1.3.5
openai==1.45.0
pinecone-client==5.0.1
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
